Consolidate timeline activity filter state
