+++ packages/twenty-front/src/modules/activities/timeline-activities/components/EventList.tsx
@@ -3,17 +3,14 @@ import { type ReactElement } from 'react';
 
 import { EventsGroup } from '@/activities/timeline-activities/components/EventsGroup';
 import { type TimelineActivity } from '@/activities/timeline-activities/types/TimelineActivity';
-import { useTimelineActivityTypes } from '@/activities/timeline-activities/hooks/useTimelineActivityTypes';
-import { timelineActivityTypeUniversalIdentifiersFilterFamilyState } from '@/activities/timeline-activities/states/timelineActivityTypeUniversalIdentifiersFilterFamilyState';
+import { useTimelineActivityTypeFilter } from '@/activities/timeline-activities/hooks/useTimelineActivityTypeFilter';
 import { filterOutInvalidTimelineActivities } from '@/activities/timeline-activities/utils/filterOutInvalidTimelineActivities';
-import { getActiveTimelineActivityTypeUniversalIdentifiersFilter } from '@/activities/timeline-activities/utils/getActiveTimelineActivityTypeUniversalIdentifiersFilter';
 import { keepTimelineActivitiesOfSelectedTypes } from '@/activities/timeline-activities/utils/keepTimelineActivitiesOfSelectedTypes';
 import { groupEventsByMonth } from '@/activities/timeline-activities/utils/groupEventsByMonth';
 import { type ActivityTargetableObject } from '@/activities/types/ActivityTargetableEntity';
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { ScrollWrapper } from '@/ui/utilities/scroll/components/ScrollWrapper';
-import { useAtomFamilyStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilyStateValue';
 import { useLingui } from '@lingui/react/macro';
 import {
   AnimatedPlaceholderEmptyContainer,
@@ -50,23 +47,10 @@ export const EventList = ({ events, targetableObject }: EventListProps) => {
 
   const { objectMetadataItems } = useObjectMetadataItems();
 
-  const { activeTimelineActivityTypes, timelineActivityTypeMaps } =
-    useTimelineActivityTypes();
-
-  const timelineActivityTypeUniversalIdentifiersFilter =
-    useAtomFamilyStateValue(
-      timelineActivityTypeUniversalIdentifiersFilterFamilyState,
-      targetableObject.id,
-    );
-
-  const effectiveTimelineActivityTypeUniversalIdentifiersFilter =
-    getActiveTimelineActivityTypeUniversalIdentifiersFilter({
-      activeUniversalIdentifiers: activeTimelineActivityTypes.map(
-        ({ universalIdentifier }) => universalIdentifier,
-      ),
-      selectedUniversalIdentifiers:
-        timelineActivityTypeUniversalIdentifiersFilter,
-    });
+  const {
+    effectiveTimelineActivityTypeUniversalIdentifiersFilter,
+    timelineActivityTypeMaps,
+  } = useTimelineActivityTypeFilter(targetableObject.id);
 
   const filteredEvents = filterOutInvalidTimelineActivities(
     keepTimelineActivitiesOfSelectedTypes(
