+++ packages/twenty-front/src/modules/activities/timeline-activities/hooks/useTimelineActivityTypeFilter.ts
@@ -1,0 +1,28 @@
+import { useTimelineActivityTypes } from '@/activities/timeline-activities/hooks/useTimelineActivityTypes';
+import { timelineActivityTypeUniversalIdentifiersFilterFamilyState } from '@/activities/timeline-activities/states/timelineActivityTypeUniversalIdentifiersFilterFamilyState';
+import { getActiveTimelineActivityTypeUniversalIdentifiersFilter } from '@/activities/timeline-activities/utils/getActiveTimelineActivityTypeUniversalIdentifiersFilter';
+import { useAtomFamilyStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilyStateValue';
+
+export const useTimelineActivityTypeFilter = (recordId: string) => {
+  const { activeTimelineActivityTypes, timelineActivityTypeMaps } =
+    useTimelineActivityTypes();
+  const selectedTimelineActivityTypeUniversalIdentifiers =
+    useAtomFamilyStateValue(
+      timelineActivityTypeUniversalIdentifiersFilterFamilyState,
+      recordId,
+    );
+
+  return {
+    activeTimelineActivityTypes,
+    effectiveTimelineActivityTypeUniversalIdentifiersFilter:
+      getActiveTimelineActivityTypeUniversalIdentifiersFilter({
+        activeUniversalIdentifiers: activeTimelineActivityTypes.map(
+          ({ universalIdentifier }) => universalIdentifier,
+        ),
+        selectedUniversalIdentifiers:
+          selectedTimelineActivityTypeUniversalIdentifiers,
+      }),
+    selectedTimelineActivityTypeUniversalIdentifiers,
+    timelineActivityTypeMaps,
+  };
+};
