+++ packages/twenty-front/src/pages/settings/applications/components/__stories__/SettingsApplicationTimelineActivityTypesListCard.stories.tsx
@@ -16,7 +16,6 @@ const meta: Meta<typeof SettingsApplicationTimelineActivityTypesListCard> = {
         icon: 'IconPaperclip',
         id: 'attachment-linked',
         isActive: true,
-        isInstalled: true,
         label: 'Attached a file',
         name: 'attachmentLinked',
       },
@@ -25,7 +24,6 @@ const meta: Meta<typeof SettingsApplicationTimelineActivityTypesListCard> = {
         icon: 'IconUnlink',
         id: 'attachment-unlinked',
         isActive: false,
-        isInstalled: true,
         label: 'Removed an attachment',
         name: 'attachmentUnlinked',
       },
