+++ packages/twenty-front/src/pages/settings/applications/hooks/useInstalledTimelineActivityTypes.ts
@@ -1,0 +1,17 @@
+import { useQuery } from '@apollo/client/react';
+import { FindManyTimelineActivityTypesDocument } from '~/generated-metadata/graphql';
+
+export const useInstalledTimelineActivityTypes = ({
+  isInstalledApplication,
+}: {
+  isInstalledApplication: boolean;
+}) => {
+  const { data, loading } = useQuery(FindManyTimelineActivityTypesDocument, {
+    skip: !isInstalledApplication,
+  });
+
+  return {
+    installedTimelineActivityTypes: data?.timelineActivityTypes ?? [],
+    loading,
+  };
+};
