+++ packages/twenty-front/src/pages/settings/applications/utils/__tests__/filterSettingsApplicationTimelineActivityTypes.test.ts
@@ -7,7 +7,6 @@ const timelineActivityTypes: SettingsApplicationTimelineActivityType[] = [
     icon: 'IconPaperclip',
     id: 'attachment-linked',
     isActive: true,
-    isInstalled: true,
     label: 'Attached a file',
     name: 'attachmentLinked',
   },
@@ -16,7 +15,6 @@ const timelineActivityTypes: SettingsApplicationTimelineActivityType[] = [
     icon: 'IconUnlink',
     id: 'attachment-unlinked',
     isActive: false,
-    isInstalled: true,
     label: 'Removed an attachment',
     name: 'attachmentUnlinked',
   },
