+++ packages/twenty-front/src/pages/settings/applications/utils/__tests__/getSettingsApplicationTimelineActivityTypes.test.ts
@@ -36,7 +36,6 @@ describe('getSettingsApplicationTimelineActivityTypes', () => {
         icon: 'IconPlus',
         id: 'type-1',
         isActive: false,
-        isInstalled: true,
         label: 'Was created',
         name: 'record.created',
       },
@@ -67,7 +66,6 @@ describe('getSettingsApplicationTimelineActivityTypes', () => {
         icon: undefined,
         id: 'universal-type-1',
         isActive: true,
-        isInstalled: false,
         label: 'Launched a rocket',
         name: 'rocket.launched',
       },
